+++ src/dsp_tools/clients/ontology_clients.py
@@ -13,7 +13,7 @@ class OntologyCreateClient(Protocol):
     """
 
     server: str
-    authentication_client: AuthenticationClient
+    auth: AuthenticationClient
 
     def get_last_modification_date(self, project_iri: str, onto_iri: str) -> Literal:
         """Get the last modification date of an ontology"""
@@ -24,7 +24,7 @@ class OntologyCreateClient(Protocol):
     def post_new_property(self, property_graph: dict[str, Any]) -> Literal | ResponseCodeAndText:
         """Create a property on the server"""
 
-    def post_new_class(self, property_graph: dict[str, Any]) -> Literal | ResponseCodeAndText:
+    def post_new_class(self, class_graph: dict[str, Any]) -> Literal | ResponseCodeAndText:
         """Create a class on the server"""
 
     def post_new_ontology(self, onto_graph: dict[str, Any]) -> str | ResponseCodeAndText: