+++ test/unittests/clients/test_metadata_client_live.py
@@ -24,7 +24,7 @@ def mock_auth_client() -> Mock:
 def metadata_client(mock_auth_client: Mock) -> MetadataClientLive:
     return MetadataClientLive(
         server="http://0.0.0.0:3333",
-        authentication_client=mock_auth_client,
+        auth=mock_auth_client,
     )